+++ saleor/graphql/product/mutations/product_variant/__init__.py
@@ -5,6 +5,7 @@ from .product_variant_reorder import ProductVariantReorder
 from .product_variant_set_default import ProductVariantSetDefault
 from .product_variant_update import ProductVariantUpdate
 from .variant_media_assign import VariantMediaAssign
+from .variant_media_reorder import VariantMediaReorder
 from .variant_media_unassign import VariantMediaUnassign
 
 __all__ = [
@@ -15,5 +16,6 @@ __all__ = [
     "ProductVariantSetDefault",
     "ProductVariantUpdate",
     "VariantMediaAssign",
+    "VariantMediaReorder",
     "VariantMediaUnassign",
 ]